+++ pvzclass/PVZ.cpp
@@ -151,9 +151,4 @@ PVZ::SaveData PVZ::GetSaveData()
 	return SaveData(Memory::ReadPointer(0x6A9EC0, 0x82C));
 }
 
-PVZ::Music PVZ::GetMusic()
-{
-	return Music(Memory::ReadPointer(0x6A9EC0, 0x83C));
-}
-
 #pragma endregion
